+++ testutil/keeper/crosschain.go
@@ -417,10 +417,6 @@ func MockVoteOnOutboundFailedBallot(
 		Once()
 }
 
-func MockGetOutbound(m *crosschainmocks.CrosschainObserverKeeper, ctx sdk.Context) {
-	m.On("GetTSS", ctx).Return(observertypes.TSS{}, true).Once()
-}
-
 func MockSaveOutbound(
 	m *crosschainmocks.CrosschainObserverKeeper,
 	ctx sdk.Context,
@@ -431,7 +427,6 @@ func MockSaveOutbound(
 	m.On("RemoveFromPendingNonces",
 		ctx, tss.TssPubkey, cctx.GetCurrentOutboundParam().ReceiverChainId, mock.Anything).
 		Return().Times(expectedNumberOfOutboundParams)
-	m.On("GetTSS", ctx).Return(observertypes.TSS{}, true)
 }
 
 func MockSaveOutboundNewRevertCreated(