+++ x/crosschain/genesis.go
@@ -43,9 +43,12 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 	}
 
 	// Set all the cross-chain txs
-	for _, elem := range genState.CrossChainTxs {
-		if elem != nil {
-			k.SetCctxAndNonceToCctxAndInboundHashToCctx(ctx, *elem)
+	tss, found := k.GetObserverKeeper().GetTSS(ctx)
+	if found {
+		for _, elem := range genState.CrossChainTxs {
+			if elem != nil {
+				k.SetCctxAndNonceToCctxAndInboundHashToCctx(ctx, *elem, tss.TssPubkey)
+			}
 		}
 	}
 	for _, elem := range genState.FinalizedInbounds {