+++ x/crosschain/keeper/grpc_query_inbound_hash_to_cctx_test.go
@@ -126,7 +126,11 @@ func TestInTxHashToCctxQueryPaginated(t *testing.T) {
 	})
 }
 
-func createInTxHashToCctxWithCctxs(keeper *crosschainkeeper.Keeper, ctx sdk.Context) ([]types.CrossChainTx,
+func createInTxHashToCctxWithCctxs(
+	ctx sdk.Context,
+	keeper *crosschainkeeper.Keeper,
+	tssPubkey string,
+) ([]types.CrossChainTx,
 	types.InboundHashToCctx) {
 	cctxs := make([]types.CrossChainTx, 5)
 	for i := range cctxs {
@@ -136,7 +140,7 @@ func createInTxHashToCctxWithCctxs(keeper *crosschainkeeper.Keeper, ctx sdk.Cont
 		cctxs[i].InboundParams = &types.InboundParams{ObservedHash: fmt.Sprintf("%d", i), Amount: math.OneUint()}
 		cctxs[i].CctxStatus = &types.Status{Status: types.CctxStatus_PendingInbound}
 		cctxs[i].RevertOptions = types.NewEmptyRevertOptions()
-		keeper.SetCctxAndNonceToCctxAndInboundHashToCctx(ctx, cctxs[i])
+		keeper.SetCctxAndNonceToCctxAndInboundHashToCctx(ctx, cctxs[i], tssPubkey)
 	}
 
 	var inboundHashToCctx types.InboundHashToCctx
@@ -152,9 +156,10 @@ func createInTxHashToCctxWithCctxs(keeper *crosschainkeeper.Keeper, ctx sdk.Cont
 func TestKeeper_InTxHashToCctxDataQuery(t *testing.T) {
 	keeper, ctx, _, zk := keepertest.CrosschainKeeper(t)
 	wctx := sdk.WrapSDKContext(ctx)
-	zk.ObserverKeeper.SetTSS(ctx, sample.Tss())
+	tss := sample.Tss()
+	zk.ObserverKeeper.SetTSS(ctx, tss)
 	t.Run("can query all cctxs data with in tx hash", func(t *testing.T) {
-		cctxs, inboundHashToCctx := createInTxHashToCctxWithCctxs(keeper, ctx)
+		cctxs, inboundHashToCctx := createInTxHashToCctxWithCctxs(ctx, keeper, tss.TssPubkey)
 		req := &types.QueryInboundHashToCctxDataRequest{
 			InboundHash: inboundHashToCctx.InboundHash,
 		}